feat: `tsci config set` && refactor(clone): improve error message, author name customization  (#183)
* refactor(clone): simplify directory path by removing author prefix

The directory path for cloned snippets no longer includes the author prefix, simplifying the structure and making it more consistent with user expectations. This change also updates the corresponding test cases to reflect the new directory naming convention.

* feat(cli): add --include-author flag to clone command

This commit introduces a new `--include-author` flag to the clone command, which includes the author's name in the directory path when cloning a snippet. This change helps avoid naming conflicts and improves project organization by clearly distinguishing projects by their authors.

* refactor(cli): simplify include-author flag and add corresponding test

The `-ia, --include-author` flag was simplified to `-a, --include-author` for better usability. Additionally, a new test was added to verify the functionality of the `-a` flag.

* feat(cli): add alwaysCloneWithAuthorName config option

Introduce a new configuration option `alwaysCloneWithAuthorName` to control whether the author name is included in the cloned snippet directory by default. This allows users to set a global preference for including the author name, reducing the need to specify the `--include-author` flag repeatedly.

Additionally, improve error handling for the clone command by providing a more user-friendly message when a snippet is not found. Register a new `config set` command to allow users to modify configuration values programmatically.